Unify `api.contact_model` in `api.contact`
