+++ src/jaxsim/api/model.py
@@ -2102,7 +2102,7 @@ def step(
 
         # Compute the 6D forces W_f ∈ ℝ^{n_L × 6} applied to links due to contact
         # with the terrain.
-        W_f_L_terrain, aux_dict = js.contact_model.link_contact_forces(
+        W_f_L_terrain, aux_dict = js.contact.link_contact_forces(
             model=model,
             data=data,
             link_forces=W_f_L_external,
